+++ src/defines.h
@@ -13,7 +13,7 @@
 //-------------------------------------
 #define VERSION_MAJOR       0
 #define VERSION_MINOR       8
-#define VERSION_PATCH       53
+#define VERSION_PATCH       54
 
 //-------------------------------------
 typedef struct {
@@ -94,7 +94,6 @@ enum {MQTT_STATUS_OFFLINE = 0, MQTT_STATUS_PARTIAL, MQTT_STATUS_ONLINE};
 
 #define MQTT_MAX_PACKET_SIZE    384
 
-#define PLUGIN_DISPLAY
 
 typedef struct {
     uint32_t rxFail;