+++ src/hm/CommQueue.h
@@ -11,9 +11,10 @@
 #include "hmInverter.h"
 #include "../utils/dbg.h"
 
-#define DEFAULT_ATTEMPS                 10
-#define MORE_ATTEMPS_ALARMDATA          15
-#define MORE_ATTEMPS_GRIDONPROFILEPARA  15
+// needs a '+1' because the comparison does not send if attempts is equal 0
+#define DEFAULT_ATTEMPS                 5 + 1
+#define MORE_ATTEMPS_ALARMDATA          15 + 1
+#define MORE_ATTEMPS_GRIDONPROFILEPARA  15 + 1
 
 template <uint8_t N=100>
 class CommQueue {